add the new tab system in MainLayout
