+++ app/src/main/AndroidManifest.xml
@@ -27,11 +27,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".LoginActivity" />
-        <activity
-            android:name=".MainActivityTabbed"
-            android:label="@string/title_activity_main_tabbed"
-            android:theme="@style/AppTheme.NoActionBar" >
-        </activity>
     </application>
 
 </manifest>
