--- app/src/main/java/com/greenfox/opal/gitinder/MainActivityTabbed.java
@@ -1,35 +1,0 @@
-package com.greenfox.opal.gitinder;
-
-import android.support.design.widget.TabLayout;
-import android.support.v7.app.AppCompatActivity;
-
-import android.support.v4.view.ViewPager;
-import android.os.Bundle;
-
-
-public class MainActivityTabbed extends AppCompatActivity {
-
-  private SectionsPagerAdapter mSectionsPagerAdapter;
-  private ViewPager mViewPager;
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_main_tabbed);
-
-    mSectionsPagerAdapter = new SectionsPagerAdapter(getSupportFragmentManager());
-    mViewPager = (ViewPager) findViewById(R.id.container);
-    setupViewPager(mViewPager);
-
-    TabLayout tabLayout = (TabLayout) findViewById(R.id.tabs);
-    tabLayout.setupWithViewPager(mViewPager);
-  }
-
-  public void setupViewPager(ViewPager viewPager) {
-    SectionsPagerAdapter adapter = new SectionsPagerAdapter(getSupportFragmentManager());
-    adapter.addFragment(new SwipingFragment(), "Swiping");
-    adapter.addFragment(new MatchesFragment(), "Matches");
-    adapter.addFragment(new SettingsFragment(), "Settings");
-    viewPager.setAdapter(adapter);
-  }
-}
