+++ app/src/main/res/menu/menu_main_activity_tabbed.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.greenfox.opal.gitinder.MainActivityTabbed" >
+    tools:context="com.greenfox.opal.gitinder.MainActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="100"
