+++ app/src/main/res/values/strings.xml
@@ -14,6 +14,5 @@
   <string name="settings_background_sync">Enable Background Sync</string>
   <string name="settings_distance">Maximum distance</string>
   <string name="settings_languages">Preferred languages</string>
-  <string name="title_activity_main_tabbed">GiTinder</string>
   <string name="action_settings">Action Settings</string>
 </resources>
